Move the preprocessing functions in dataset.py.
